Rename converter directory to endpoint-converter
