+++ src/core/interfaces-writer/UrlStructureInterfacesWriter.ts
@@ -1,10 +1,10 @@
-import {InterfacesWriter} from "./InterfacesWriter";
-import {ConverterResult} from "../converter/ApiDocToInterfaceConverter";
-import {ApiDoc2InterfaceParameters} from "../ApiDoc2Interface";
-import {getUrlWithoutParameters} from "../StringUtils";
-import {writeFileToPath} from "../FsUtils";
-import {stringifyInterfaces} from "./WriterUtils";
 import * as path from "path";
+import {ApiDoc2InterfaceParameters} from "../ApiDoc2Interface";
+import {ConverterResult} from "../endpoint-converter/ApiDocToInterfaceConverter";
+import {writeFileToPath} from "../FsUtils";
+import {getUrlWithoutParameters} from "../StringUtils";
+import {InterfacesWriter} from "./InterfacesWriter";
+import {stringifyInterfaces} from "./WriterUtils";
 
 export class UrlStructureInterfacesWriter implements InterfacesWriter {
     async writeInterfaces(interfacesData: Array<ConverterResult>, args: ApiDoc2InterfaceParameters): Promise<void> {
